feat(plasma-new-hope): rename DatePicker folders
