+++ packages/plasma-new-hope/src/components/Calendar/ui/CalendarYears/CalendarYears.styles.ts
@@ -60,7 +60,7 @@ export const StyledYearRoot = styled.div`
             outlineSize: '0.063rem',
             outlineOffset: '-0.1875rem',
             outlineColor: `var(${tokens.calendarOutlineFocusColor})`,
-            outlineRadius: `calc(var(${tokens.calendarYearItemBorderRadius}) - 0.125rem)`,
+            outlineRadius: `calc(var(${tokens.calendarYearItemBorderRadius}) + 0.125rem)`,
         })};
     }
 `;
