+++ packages/plasma-new-hope/src/components/DatePicker/index.ts
@@ -1,7 +1,7 @@
-export { datePickerRoot, datePickerConfig } from './SingleDate/DatePicker';
-export type { DatePickerProps, DatePickerPlacementBasic, DatePickerPlacement } from './SingleDate/DatePicker.types';
+export { datePickerRoot, datePickerConfig } from './SingleDate/SingleDate';
+export type { DatePickerProps, DatePickerPlacementBasic, DatePickerPlacement } from './SingleDate/SingleDate.types';
 
-export { datePickerRangeRoot, datePickerRangeConfig } from './RangeDate/DatePicker';
-export type { DatePickerRangeProps, DatePickerRangePlacement } from './RangeDate/DatePicker.types';
+export { datePickerRangeRoot, datePickerRangeConfig } from './RangeDate/RangeDate';
+export type { DatePickerRangeProps, DatePickerRangePlacement } from './RangeDate/RangeDate.types';
 
 export { classes as datePickerClasses, tokens as datePickerTokens } from './DatePicker.tokens';
